feat: Establish all logic for slider and puzzle
I have created the logic for the puzzle and slider to function
properly. I have also included the 15 puzzle as well.
You will notice that I wrote the code so as to not have to refactor for
a fdifferent size puzzle. Styling hold up on both all the way to highest zoom.
in the slider it is also fully functional and responsive.